Refine solutions for ay23/24 s2

+++ optional 2/WiFi.java
@@ -11,8 +11,6 @@ class WiFi {
         double end = (houses[houses.length - 1] - houses[0]) / 2.0;
         while (end - begin >= 0.5) {
             double mid = (begin + end) / 2;
-            // if I write mid = (begin + end) / 2.0, the code fails one of the private test cases
-            // and I don't know why
             if (coverable(houses, numOfAccessPoints, mid)) {
                 end = mid;
             } else {
@@ -29,7 +27,7 @@ class WiFi {
         Arrays.sort(houses);
         int n = houses.length;
         int accessPointCount = 1;
-        int currentLength = 0;
+        double currentLength = 0;
         for (int i = 1; i < n; i++) {
             currentLength -= houses[i - 1] - houses[i];
             if (currentLength > 2 * distance) {
